get today note

+++ src/services/DoneService.ts
@@ -1,5 +1,6 @@
 import { DoneModel } from "../models/DoneModel";
 import { DoneRepository } from "../repositories/DoneRepository";
+import { validateCreateDate } from "../utils/validateCreateDate";
 
 class DoneService {
 
@@ -10,6 +11,10 @@ class DoneService {
 	}
 
 	async createDone(userId: string, done: DoneModel) {
+		const note = await this.repository.findNote(userId, done.noteId)
+		if (!note) throw new Error("Id inválido");
+		validateCreateDate(note.date)
+
 		if (done.minutes < 0 || done.minutes > (60 * 24)) throw new Error("Minutos inválidos");
 		if (await this.repository.findNote(userId, done.noteId)) {
 			await this.repository.createDone(done)
@@ -19,6 +24,10 @@ class DoneService {
 	}
 
 	async updateDone(userId: string, done: DoneModel) {
+		const note = await this.repository.findNote(userId, done.noteId)
+		if (!note) throw new Error("Id inválido");
+		validateCreateDate(note.date)
+
 		if (done.minutes < 0 || done.minutes > (60 * 24)) throw new Error("Minutos inválidos");
 		if (await this.repository.findNote(userId, done.noteId)) {
 			await this.repository.updateDone(done)
@@ -28,7 +37,12 @@ class DoneService {
 	}
 
 	async deleteDone(userId: string, doneId: string | null) {
-		if(doneId == null) throw new Error("Id inválido");
+		if (!doneId) throw new Error("Id inválido");
+		const note = await this.repository.findNoteByDone(doneId)
+		if (!note) throw new Error("Id inválido");
+		validateCreateDate(note.date)
+
+		if (doneId == null) throw new Error("Id inválido");
 		await this.repository.deleteDone(userId, doneId)
 	}
 }
